benchmarks: split agent-workflow audits into per-task leaves
Remove the shared test_task_regressions_audits.py dump and the generic
corrupted-witness mutate table so Harbor task PRs stop colliding on one
file. Keep a small fixed RESULT_JSON policy sample and document the
per-task leaf convention.
