+++ benchmarks/validation/agent_workflow_v1/test_autoformalization_semantic_audit.py
@@ -1,0 +1,34 @@
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+from benchmarks.validation.agent_workflow_v1 import support
+
+TASK = "autoformalization-semantic-audit"
+
+
+def test_rejects_positive_lean_compile_claim(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    (app / "evidence" / "answer.txt").write_text(
+        "dimension dot product coordinate\n"
+        "Both Lean declarations compile.\n"
+        "RESULT_JSON: {}\n"
+    )
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 1.0
+    assert rejected["evidence_validity"] == 0.0
+
+
+def test_rejects_corrupted_operator_mismatch(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["operator_mismatch_certificate"]["dot_product"] = 1
+    support._write_json(app / "submission.json", submission)
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
+    assert rejected["reward"] == 0.0
