+++ benchmarks/validation/agent_workflow_v1/test_divisibility_construction_witness.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+from benchmarks.validation.agent_workflow_v1 import support
+
+TASK = "divisibility-construction-witness"
+
+
+def test_accepts_schema_valid_integral_numbers(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"] = {
+        key: float(value) for key, value in submission["result"].items()
+    }
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+    accepted = support._run_verifier(task, app, logs)
+    assert accepted["correctness"] == 1.0
