+++ benchmarks/validation/agent_workflow_v1/test_gaussian_moment_generality_audit.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+from benchmarks.validation.agent_workflow_v1 import support
+
+TASK = "gaussian-moment-generality-audit"
+
+
+def test_rejects_corrupted_coefficient(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["v_coefficients"][2] = "-1/3"
+    support._write_json(app / "submission.json", submission)
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
+    assert rejected["reward"] == 0.0
