+++ benchmarks/validation/agent_workflow_v1/test_infinite_shift_spectrum_counterexample.py
@@ -1,0 +1,51 @@
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+from benchmarks.validation.agent_workflow_v1 import support
+
+TASK = "infinite-shift-spectrum-counterexample"
+
+
+def _case(tmp_path: Path):
+    return support._prepare_case(tmp_path, TASK, "computed")
+
+
+def _rewrite(app: Path, submission: dict) -> None:
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+
+def test_accepts_reversed_operator_orientation(tmp_path: Path) -> None:
+    task, app, logs = _case(tmp_path)
+    submission = json.loads((app / "submission.json").read_text())
+    result = submission["result"]
+    result["orientation"] = "S_LEFT_T_RIGHT"
+    result["zero_eigenvalue_product"] = "TS"
+    result["identity_product"] = "ST"
+    for action in result["actions"]:
+        index = action["basis_index"]
+        action.update(
+            {
+                "s_output": None if index == 0 else index - 1,
+                "t_output": index + 1,
+                "st_output": index,
+                "ts_output": None if index == 0 else index,
+            }
+        )
+    _rewrite(app, submission)
+
+    accepted = support._run_verifier(task, app, logs)
+    assert accepted["correctness"] == 1.0
+
+
+def test_rejects_corrupted_composition(tmp_path: Path) -> None:
+    task, app, logs = _case(tmp_path)
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["actions"][5]["st_output"] = 4
+    _rewrite(app, submission)
+
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
+    assert rejected["reward"] == 0.0
