+++ benchmarks/validation/agent_workflow_v1/test_matrix_square_zero_counterexample.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+from benchmarks.validation.agent_workflow_v1 import support
+
+TASK = "matrix-square-zero-counterexample"
+
+
+def test_rejects_corrupted_matrix(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["matrix"] = [[1, 0], [0, 0]]
+    support._write_json(app / "submission.json", submission)
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
+    assert rejected["reward"] == 0.0
